+++ wrappers/go/zxc_seekable.go
@@ -183,11 +183,13 @@ func (s *Seekable) Close() error {
 // NumBlocks returns the total number of data blocks in the archive
 // (excluding the EOF marker block). Returns 0 if the handle has been
 // closed.
-func (s *Seekable) NumBlocks() uint32 {
+//
+// Derived from the footer, never stored, so only the archive size bounds it.
+func (s *Seekable) NumBlocks() uint64 {
 	if s == nil || s.ptr == nil {
 		return 0
 	}
-	return uint32(C.zxc_seekable_get_num_blocks(s.ptr))
+	return uint64(C.zxc_seekable_get_num_blocks(s.ptr))
 }
 
 // DecompressedSize returns the total decompressed size of the archive in
@@ -204,12 +206,12 @@ func (s *Seekable) DecompressedSize() uint64 {
 // group and checked against bounds only. ok is false if blockIdx is out of
 // range or the handle is closed; err is [ErrInvalidData] if the group is
 // unreadable or invalid.
-func (s *Seekable) BlockCompressedSize(blockIdx uint32) (size uint32, ok bool, err error) {
+func (s *Seekable) BlockCompressedSize(blockIdx uint64) (size uint32, ok bool, err error) {
 	if s == nil || s.ptr == nil || blockIdx >= s.NumBlocks() {
 		return 0, false, nil
 	}
 	s.busy++
-	sz := uint32(C.zxc_seekable_get_block_comp_size(s.ptr, C.uint32_t(blockIdx)))
+	sz := uint32(C.zxc_seekable_get_block_comp_size(s.ptr, C.uint64_t(blockIdx)))
 	s.busy--
 	// 0 is never a real size: the group is unreadable or invalid.
 	if sz == 0 {
@@ -221,11 +223,11 @@ func (s *Seekable) BlockCompressedSize(blockIdx uint32) (size uint32, ok bool, e
 // BlockDecompressedSize returns the decompressed size of a specific block.
 // The second return value is false if blockIdx is out of range or the
 // handle has been closed.
-func (s *Seekable) BlockDecompressedSize(blockIdx uint32) (uint32, bool) {
+func (s *Seekable) BlockDecompressedSize(blockIdx uint64) (uint32, bool) {
 	if s == nil || s.ptr == nil || blockIdx >= s.NumBlocks() {
 		return 0, false
 	}
-	return uint32(C.zxc_seekable_get_block_decomp_size(s.ptr, C.uint32_t(blockIdx))), true
+	return uint32(C.zxc_seekable_get_block_decomp_size(s.ptr, C.uint64_t(blockIdx))), true
 }
 
 // DecompressRange decompresses length bytes starting at offset (in the
@@ -336,9 +338,9 @@ func (s *Seekable) SetDict(dict, hufLengths []byte) error {
 
 // SeekTableSize returns the encoded byte size of a seek table covering
 // numBlocks data blocks. Use this to size a destination buffer for
-// [WriteSeekTable].
-func SeekTableSize(numBlocks uint32) int {
-	return int(C.zxc_seek_table_size(C.uint32_t(numBlocks)))
+// [WriteSeekTable]. Returns 0 when that size does not fit a size_t.
+func SeekTableSize(numBlocks uint64) int {
+	return int(C.zxc_seek_table_size(C.uint64_t(numBlocks)))
 }
 
 // WriteSeekTable writes a seek table (header + entries) into dst.
@@ -357,7 +359,7 @@ func WriteSeekTable(dst []byte, compSizes []uint32) (int, error) {
 		(*C.uint8_t)(unsafe.Pointer(&dst[0])),
 		C.size_t(len(dst)),
 		(*C.uint32_t)(unsafe.Pointer(&compSizes[0])),
-		C.uint32_t(len(compSizes)),
+		C.uint64_t(len(compSizes)),
 	)
 	if res < 0 {
 		return 0, errorFromCode(res)
